feat: MovingBusAbsolute -> MovingBus
